Update Merge-MailboxFolder.ps1
I am slowly adding support for retention policy support for when you migrate an archive back into a users mailbox. Retention changed in 2013 and EWS needs to have a higher versions used to have access to those properties.  I updated this Parameter to be a little more flexible with specifying other version of Exchange.